requirements.txt
requests
pandas
bs4
playwright
html2text
python-dotenv
openpyxl
langchain_openai
tqdm
streamlit
dnspython
